feat: taxi api done (#18)
* taxi_update

yes

* TaxiAPI_updated functions

Added the respective functions that returns a map of the price and name of the taxi service

* Update taxi.tsx

* TAXI_API_changes

changes to the 3 functions. fixed export default class Taxi

* update taxi.tsx

kjxkvhxlkdfjglk

* Update taxi.tsx

finalized taxi.tsx file (accounts for all pax num, etc.)

* taxi_update

yes

* TaxiAPI_updated functions

Added the respective functions that returns a map of the price and name of the taxi service

* Update taxi.tsx

* TAXI_API_changes

changes to the 3 functions. fixed export default class Taxi

* fixing taxi API

* fix: remove git merge syntax

* fix: remove unnecessary import

---------

Authored-by: teg tiwana <82508380+tej172@users.noreply.github.com>
Co-authored-by: joanne-ch <joanne.christina6@gmail.com>